requirements.txt
# Python dependencies for Twitter AI Automation Tool
# Install with: pip install -r requirements.txt

tweepy==4.14.0
google-genai==0.2.0
schedule==1.2.0
python-dotenv==1.0.0